Working on adding functionality and some cleanup
