+++ src/main/java/Main.java
@@ -1,14 +1,14 @@
-import view.add_connection;
+import view.Add_connection;
 
 import javax.swing.*;
 
-public class main {
+public class Main {
     public static void main(String[] args)
     {
         SwingUtilities.invokeLater(new Runnable() {
             @Override
             public void run() {
-                add_connection conn = new add_connection();
+                Add_connection conn = new Add_connection();
                 conn.newcon();
             }
         });
