+++ src/main/java/controller/Connect.java
@@ -2,11 +2,10 @@ package controller;
 import model.Connection;
 
 import com.jcraft.jsch.*;
-import model.Connection;
 
-public class connect {
+public class Connect {
 
-    public void connect(Connection connection) {
+    public void openConnection(Connection connection) {
         JSch jsch = new JSch();
         try {
             // Args are username, pw, ip
